# docs/openapi.yaml
openapi: 3.0.3
info:
  title: sandboxd /v1 API
  version: "0.3.0"
  description: >
    The public, tenant-scoped API for sandboxd. This is the only surface a
    console or external integration should use — the internal /sandbox and
    /wake endpoints are not part of this contract. Scoped by the API token
    (Bearer); when auth is disabled (the single-user default) the whole
    surface is reachable without a token.
servers:
  - url: http://127.0.0.1:9090
    description: Default loopback bind (docker-compose maps SANDBOXD_API_BIND here)
security:
  - bearerAuth: []
  - {} # auth disabled by default (single-user)
...
paths:
...
  /v1/apps/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: getApp
      tags: [apps]
      summary: Get an app
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/App" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateApp
      tags: [apps]
      summary: Update an app (partial)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchAppRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/App" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteApp
      tags: [apps]
      summary: Delete an app and EVERYTHING it owns — purges every sandbox (container, workspace image, snapshots) and all app rows (config, events, snapshots). Irreversible.
      responses:
        "204": { description: App and all its resources deleted }
        "404": { $ref: "#/components/responses/Error" }
...
  /v1/apps/{id}/snapshots:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: listAppSnapshots
      tags: [apps]
      summary: List snapshots captured from this app (tenant-scoped, newest first)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  snapshots:
                    type: array
                    items: { $ref: "#/components/schemas/Snapshot" }
        "404": { $ref: "#/components/responses/Error" }
  /v1/apps/{id}/restore:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      operationId: restoreApp
      tags: [apps]
      summary: Restore the app source from a snapshot
      description: >
        Un-snapshotted source changes are lost, so the console confirms before
        calling. Cube imports source inside the existing VM and retains its
        private home, databases, credentials and sandbox ID. A different runtime
        template requires explicit migration first. Docker retains its legacy
        destructive replacement behavior (container and workspace). The response
        reports current preview/process readiness; source acknowledgement alone
        does not mean application or database startup has completed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [snapshot_id]
              properties:
                snapshot_id: { type: string }
      responses:
        "201":
          description: Source restored; existing Cube sandbox retained or new sandbox created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sandbox" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /v1/apps/{id}/fork:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      operationId: forkApp
      tags: [apps]
      summary: Fork a snapshot into a new app (source app untouched)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [snapshot_id]
              properties:
                snapshot_id: { type: string }
                name: { type: string, description: "New app name; defaults to '<source> (fork)'." }
                external_user_id: { type: string, description: "Ownership override for the new app; empty inherits the source app's." }
                external_project_id: { type: string, description: "External project id override for the new app; empty inherits the source app's." }
      responses:
        "201":
          description: New app, with its sandbox spun from the snapshot
          content:
            application/json:
              schema:
                type: object
                properties:
                  app: { $ref: "#/components/schemas/App" }
                  sandbox: { $ref: "#/components/schemas/Sandbox" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
  /v1/apps/{id}/config:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: listAppConfig
      tags: [apps]
      summary: List an app's config entries (metadata; secret values redacted)
      description: >
        Returns every config key for the app. Sensitive entries never
        include their value (only value_set=true); non-sensitive entries
        include the plaintext value.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  config:
                    type: array
                    items: { $ref: "#/components/schemas/ConfigItem" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      operationId: createAppConfig
      tags: [apps]
      summary: Create a config entry (set sensitive=true to encrypt at rest)
      description: >
        Sensitive values are AES-256-GCM-encrypted at rest and are
        redacted — they are never returned by any GET. access_policy
        governs who may later read the value through the broker; the
        default control_plane_only is not supplied to runtime processes.
        The owning tenant service may explicitly reveal one stored value
        using the separate POST reveal endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateConfigRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigItem" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: A config entry with this key already exists; PATCH to update.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/apps/{id}/config/{key}:
    parameters:
      - $ref: "#/components/parameters/Id"
      - in: path
        name: key
        required: true
        schema: { type: string }
    patch:
      operationId: updateAppConfig
      tags: [apps]
      summary: Update a config entry (value, sensitivity, or access policy)
      description: >
        Omit value to keep the stored value untouched (a policy-only
        change preserves an existing secret). Toggling sensitive
        re-encodes the stored value.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchConfigRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigItem" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteAppConfig
      tags: [apps]
      summary: Delete a config entry
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/Error" }
  /v1/apps/{id}/config/{key}/reveal:
    parameters:
      - $ref: "#/components/parameters/Id"
      - in: path
        name: key
        required: true
        schema: { type: string }
    post:
      operationId: revealAppConfig
      tags: [apps]
      summary: Explicitly reveal one key to the owning tenant service
      description: >
        Requires an authenticated service belonging to the app tenant.
        Responses are private and no-store. Audit records contain the key
        name only, never its value. GET/list remains redacted. Integrators
        must additionally authorize the individual project owner.
      responses:
        "200":
          description: Revealed value; must not be logged or cached
          content:
            application/json:
              schema:
                type: object
                required: [key, value]
                properties:
                  key: { type: string }
                  value: { type: string }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
...
  /v1/sandboxes/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: getSandbox
      tags: [sandboxes]
      summary: Get a sandbox (status + live preview state)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sandbox" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteSandbox
      tags: [sandboxes]
      summary: Destroy the sandbox (container + workspace)
      responses:
        "200": { description: Destroyed }
        "404": { $ref: "#/components/responses/Error" }
  /v1/sandboxes/{id}/start:
    parameters: [ { $ref: "#/components/parameters/Id" } ]
    post:
      operationId: startSandbox
      tags: [sandboxes]
      summary: Start (wake) a stopped sandbox. Idempotent when running.
      responses:
        "200":
          description: Running
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sandbox" }
        "409": { $ref: "#/components/responses/Error" }
        "503":
          description: Capacity pushback (Retry-After).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/sandboxes/{id}/preview-access:
    parameters: [ { $ref: "#/components/parameters/Id" } ]
    post:
      operationId: issueCubePreviewAccess
      tags: [sandboxes]
      summary: Issue a scoped five-minute Cube preview capability
      description: >-
        Requires the sandbox owner's service tenant. The calling platform must
        authorize its viewer before requesting or returning this capability.
        Does not wake the sandbox. Responses must not be cached. Existing
        capabilities remain valid until expiry after visibility changes.
        The browser access URL installs a host-only HttpOnly cookie and redirects
        to the clean preview URL. Third-party cookie restrictions may require
        opening the preview in a top-level tab. The token must not be included in normal sandbox responses.
      responses:
        "200":
          description: Scoped access; Cache-Control no-store
          content:
            application/json:
              schema:
                type: object
                required: [url, access_url, token, expires_at]
                properties:
                  url: { type: string, format: uri }
                  access_url: { type: string, format: uri }
                  token: { type: string }
                  expires_at: { type: string, format: date-time }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /v1/sandboxes/{id}/recreate:
    parameters: [ { $ref: "#/components/parameters/Id" } ]
    post:
      operationId: recreateSandbox
      tags: [sandboxes]
      summary: >-
        Stop, remove and rebuild the container from the stored row with the
        current image and the app's current config (delivered as environment).
        The workspace is untouched. Use after changing app config values.
        Cube applies the current config in the existing VM; an already applied
        revision is an idempotent success without another restart. With
        reload_manifest=true, Cube validates sandbox.yaml and activates its
        current worker/web commands in the same VM. Repeating an identical
        manifest/config is idempotent. The exposed web port cannot change here.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reload_manifest: { type: boolean, default: false }
      responses:
        "200":
          description: Configuration acknowledged; preview may still be starting
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sandbox" }
        "409":
          description: A task is in progress; cancel it first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Manifest invalid or exposed web port changed
        "503":
          description: Capacity pushback (Retry-After).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/sandboxes/{id}/stop:
    parameters: [ { $ref: "#/components/parameters/Id" } ]
    post:
      operationId: stopSandbox
      tags: [sandboxes]
      summary: Stop a running sandbox (frees RAM; wakes on start). Idempotent when stopped.
      responses:
        "200":
          description: Stopped
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sandbox" }
        "409":
          description: A task is in progress; cancel it first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/sandboxes/{id}/tasks:
    parameters: [ { $ref: "#/components/parameters/Id" } ]
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List a sandbox's task history (newest first)
      description: >
        Durable task history from the store — each row includes files_changed
        and can_revert (whether a pre-task checkpoint exists to go back to).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        prompt: { type: string }
                        agent: { type: string }
                        status: { type: string }
                        agent_message: { type: string }
                        files_changed: { type: array, items: { type: string } }
                        checkpoint_id: { type: string }
                        can_revert: { type: boolean }
                        created_at: { type: string, format: date-time }
    post:
      operationId: submitTask
      tags: [tasks]
      summary: Submit a coding-agent task (wakes a stopped sandbox first)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubmitTaskRequest" }
      responses:
        "202":
          description: Accepted (task running)
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  sandbox_id: { type: string }
                  status: { type: string, example: running }
                  agent: { type: string }
                  events_url: { type: string }
        "400": { $ref: "#/components/responses/Error" }
        "409":
          description: A task is already in progress.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/sandboxes/{id}/tasks/{taskId}:
    parameters:
      - $ref: "#/components/parameters/Id"
      - in: path
        name: taskId
        required: true
        schema: { type: string }
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task's canonical result
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskResult" }
        "404": { $ref: "#/components/responses/Error" }
  /v1/sandboxes/{id}/tasks/{taskId}/events:
    parameters:
      - $ref: "#/components/parameters/Id"
      - in: path
        name: taskId
        required: true
        schema: { type: string }
    get:
      operationId: streamSandboxTaskEvents
      tags: [tasks]
      summary: Live task event stream (Server-Sent Events)
      description: >
        text/event-stream of status/message/tool/build/done events. The
        console consumes this with EventSource for live status + logs.
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema: { type: string }
  /v1/sandboxes/{id}/tasks/{taskId}/cancel:
    parameters:
      - $ref: "#/components/parameters/Id"
      - in: path
        name: taskId
        required: true
        schema: { type: string }
    post:
      operationId: cancelTask
      tags: [tasks]
      summary: Cancel a running task
      responses:
        "200": { description: Cancelled }
  /v1/sandboxes/{id}/tasks/{taskId}/messages:
    parameters:
      - $ref: "#/components/parameters/Id"
      - in: path
        name: taskId
        required: true
        schema: { type: string }
    post:
      operationId: sendTaskMessage
      tags: [tasks]
      summary: Send a follow-up into the current coding-agent session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message_id, prompt]
              properties:
                message_id: { type: string, format: uuid }
                prompt: { type: string, maxLength: 80000 }
      responses:
        "202": { description: Accepted by the running session; input events acknowledge consumption }
        "400": { description: Invalid input }
        "404": { description: Unknown task or project outside this tenant }
        "409": { description: Agent is finishing or live input is unsupported }
        "502": { description: Agent could not be reached }
  /v1/sandboxes/{id}/tasks/{taskId}/revert:
    parameters:
      - $ref: "#/components/parameters/Id"
      - in: path
        name: taskId
        required: true
        schema: { type: string }
    post:
      operationId: revertTask
      tags: [tasks]
      summary: Revert the workspace to a task's pre-task checkpoint
      description: >
        Restores the working files to how they were before the task ran (undoing
        that task and everything after it), without moving the branch/HEAD.
        Requires a running sandbox; the task must have a checkpoint.
      responses:
        "200": { description: Reverted }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    cubeRelayKey:
      type: apiKey
      in: header
      name: x-api-key
    cubeBridgeToken:
      type: apiKey
      in: header
      name: x-baarcha-bridge
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    App:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        tags: { type: array, items: { type: string } }
        external_user_id: { type: string }
        external_project_id: { type: string }
        runtime_preset: { type: string, description: "Default runtime preset id for new sandboxes." }
        current_sandbox_id: { type: string, description: "Empty when the app has no live sandbox." }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    PatchAppRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        tags: { type: array, items: { type: string } }
...